+++ src/app/app/dashboard/page.tsx
@@ -7,7 +7,7 @@ import '@mux/mux-video';
 import Link from 'next/link'
 import { NotionRenderer, createBlockRenderer } from "@notion-render/client"
 
-import { Globals, Nav, Typography, TypographyVariant, ENavControlVariant, EIcon, AudioPlayer } from '@dreampipcom/oneiros'
+import { Typography, TypographyVariant, ENavControlVariant, EIcon, AudioPlayer } from '@dreampipcom/oneiros'
 import "@dreampipcom/oneiros/styles"
 
 import Layout from './layout'
@@ -36,21 +36,11 @@ export default function Template({ title, content, isomorphicContent }: any) {
   }
 
     return (
-      <Globals theme={globalContext.theme}>
-      <Nav onThemeChange={handleThemeChange} />
       <main className="min-h-[100vh]">
       <ViewMenu active="dashboard" />
       <h1 className="scroll-m-20 text-2xl font-semibold tracking-tight text-center mb-8">{new Date().toLocaleString("en-US", {weekday: "long", year: "numeric", month: "short", day: "numeric" })}</h1>
       <h2 className="text-center scroll-m-20 text-lg font-semibold tracking-tight">Your life at a glimpse.</h2>
       <AnalyticsView />
-      <footer>
-            <div className="flex w-full flex-center justify-center p-a2">
-              <Typography>
-                © 1992—Present Angelo Reale
-              </Typography>
-            </div>
-          </footer>
       </main>
-      </Globals>
     )
 }
