+++ src/app/app/week/page.tsx
@@ -7,7 +7,7 @@ import '@mux/mux-video';
 import Link from 'next/link'
 import { NotionRenderer, createBlockRenderer } from "@notion-render/client"
 
-import { Globals, Nav, Typography, TypographyVariant, ENavControlVariant, EIcon, AudioPlayer } from '@dreampipcom/oneiros'
+import { Typography, TypographyVariant, ENavControlVariant, EIcon, AudioPlayer } from '@dreampipcom/oneiros'
 import "@dreampipcom/oneiros/styles"
 
 import Layout from './layout'
@@ -58,8 +58,6 @@ export default function Template({ title, content, isomorphicContent }: any) {
     })
 
     return (
-      <Globals theme={globalContext.theme}>
-      <Nav onThemeChange={handleThemeChange} />
       <main className="min-h-[100vh]">
 
         <ViewMenu active="week" />
@@ -69,15 +67,7 @@ export default function Template({ title, content, isomorphicContent }: any) {
         <p className="text-center scroll-m-20 text-lg font-semibold tracking-tight mb-8">What did you accomplish so far?</p>
 
       <TaskView timeframe="week" actions={WEEKLY_ACTIONS} />
-      
-      <footer>
-            <div className="flex w-full flex-center justify-center p-a2">
-              <Typography>
-                © 1992—Present Angelo Reale
-              </Typography>
-            </div>
-          </footer>
+    
       </main>
-      </Globals>
     )
 }
