+++ src/app/layout.tsx
@@ -1,7 +1,10 @@
 'use client'
+import { useState } from 'react'
 import type { Metadata } from "next"
 import { Comfortaa } from "next/font/google"
 import { SessionProvider } from "next-auth/react"
+
+import { Nav, Globals } from '@dreampipcom/oneiros'
 
 import "@dreampipcom/oneiros/styles"
 import "./globals.css"
@@ -21,15 +24,39 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const [globalContext, setGlobalContext] = useState({ theme: "light" })
+
+  const handleThemeChange = () => {
+    if (globalContext.theme === 'light') {
+      setGlobalContext({...globalContext, theme: 'dark'})
+    } else {
+      setGlobalContext({...globalContext, theme: 'light'})
+    }
+  }
+
   return (
     <html lang="en">
       <meta name="viewport" content="width=device-width, initial-scale=1" />
       <body
         className={`${comfortaa.variable} antialiased`}
       >
+        
+        <Globals theme={globalContext.theme}>
+        <Nav onThemeChange={handleThemeChange} />
         <SessionProvider>
-        {children}
+          {children}
         </SessionProvider>
+      </Globals>
+        <footer>
+            <div className="flex w-full flex-center 
+              flex-col justify-center p-a2 bg-[#f1cfff]">
+              <small>
+                © 2012—Present DreamPip
+              </small>
+              <a href="/terms" className="text-sm">Terms of Service</a>
+              <a href="/privacy" className="text-sm">Privacy Policy</a>
+            </div>
+          </footer>
       </body>
     </html>
   );
